Dev: Improved Socket Error Handling

+++ src/DotnetCat/Utils/Platform.cs
@@ -3,7 +3,7 @@ namespace DotnetCat.Utils;
 /// <summary>
 ///  Operating system platform enumeration type.
 /// </summary>
-internal enum Platform : ushort
+internal enum Platform : byte
 {
     Nix,  // Linux/Unix based operating system
     Win   // Windows based operating system
